[AIP-21] Move contrib.operators.spark_submit_operator providers.apache.spark.operators.spark_submit
