+++ scripts/ci/pylint_todo.txt
@@ -68,7 +68,7 @@
 ./airflow/contrib/operators/snowflake_operator.py
 ./airflow/providers/apache/spark/operators/spark_jdbc.py
 ./airflow/providers/apache/spark/operators/spark_sql.py
-./airflow/contrib/operators/spark_submit_operator.py
+./airflow/providers/apache/spark/operators/spark_submit.py
 ./airflow/contrib/operators/sqoop_operator.py
 ./airflow/contrib/operators/ssh_operator.py
 ./airflow/contrib/operators/vertica_operator.py
