+++ tests/providers/apache/spark/operators/test_spark_submit.py
@@ -22,8 +22,8 @@ import unittest
 from datetime import timedelta
 
 from airflow import DAG
-from airflow.contrib.operators.spark_submit_operator import SparkSubmitOperator
 from airflow.models import TaskInstance
+from airflow.providers.apache.spark.operators.spark_submit import SparkSubmitOperator
 from airflow.utils import timezone
 
 DEFAULT_DATE = timezone.datetime(2017, 1, 1)
